requirements.txt
fastapi
uvicorn
pydantic
pyyaml
rich
orjson
requests
core